+++ src/basic/components/ui/SearchBar.tsx
@@ -1,0 +1,17 @@
+interface SearchBarProps {
+  value: string;
+  onChange: (e: React.ChangeEvent<HTMLInputElement>) => void;
+  placeholder?: string;
+}
+
+export function SearchBar({ value, onChange, placeholder = '상품 검색...' }: SearchBarProps) {
+  return (
+    <input
+      type="text"
+      value={value}
+      onChange={onChange}
+      placeholder={placeholder}
+      className="w-full max-w-md px-4 py-2 border border-gray-300 rounded-lg focus:outline-none focus:border-blue-500"
+    />
+  );
+}
