+++ src/basic/components/ui/index.ts
@@ -2,3 +2,4 @@ export * from './EmptyState';
 export * from './ProductList';
 export * from './CartList';
 export * from './NotificationList';
+export * from './SearchBar';
