+++ src/basic/hooks/admin/index.ts
@@ -1,1 +1,2 @@
-export * from './useAdminForm';
+export * from './useProductList';
+export * from './useCouponList';
