+++ src/basic/hooks/admin/useProductList.ts
@@ -1,7 +1,6 @@
 import { useState } from 'react';
-import { Coupon, ProductWithUI } from '../../../types';
+import { ProductWithUI } from '../../../types';
 import {
-  formatAddMessage,
   formatErrorMessageCRUD,
   formatErrorMessageProduct,
   formatExceedErrorMessage,
@@ -16,29 +15,19 @@ type ProductForm = {
   discounts: Array<{ quantity: number; rate: number }>;
 };
 
-type CouponForm = {
-  name: string;
-  code: string;
-  discountType: 'amount' | 'percentage';
-  discountValue: number;
-};
-
 type ValidationResult = {
   isValid: boolean;
   message?: string;
 };
 
-export function useAdminForm({
+export function useProductList({
   addProduct,
   updateProduct,
-  addCoupon,
 }: {
   addProduct: (product: Omit<ProductWithUI, 'id'>) => void;
   updateProduct: (id: string, updates: Partial<ProductWithUI>) => void;
-  addCoupon: (coupon: Coupon) => void;
 }) {
   const { isOpen: showProductForm, toggle: toggleProductForm } = useToggle(false);
-  const { isOpen: showCouponForm, toggle: toggleCouponForm } = useToggle(false);
 
   const [editingProduct, setEditingProduct] = useState<string | null>(null);
   const [productForm, setProductForm] = useState<ProductForm>({
@@ -47,13 +36,6 @@ export function useAdminForm({
     stock: 0,
     description: '',
     discounts: [],
-  });
-
-  const [couponForm, setCouponForm] = useState<CouponForm>({
-    name: '',
-    code: '',
-    discountType: 'amount',
-    discountValue: 0,
   });
 
   const handleProductSubmit = (e: React.FormEvent): { success: boolean; message?: string } => {
@@ -127,45 +109,7 @@ export function useAdminForm({
     toggleProductForm();
   };
 
-  // Coupon form handlers
-  const handleCouponSubmit = (e: React.FormEvent): { success: boolean; message?: string } => {
-    e.preventDefault();
-
-    // Validate coupon form before submit
-    const discountValidation = validateCouponForm('discountValue', couponForm.discountValue);
-    if (!discountValidation.isValid) {
-      return { success: false, message: discountValidation.message };
-    }
-
-    // Check required fields
-    if (!couponForm.name.trim()) {
-      return { success: false, message: '쿠폰명을 입력해주세요.' };
-    }
-
-    if (!couponForm.code.trim()) {
-      return { success: false, message: '쿠폰 코드를 입력해주세요.' };
-    }
-
-    try {
-      const newCoupon = couponForm as Coupon;
-      addCoupon(newCoupon);
-      return { success: true, message: formatAddMessage('쿠폰') };
-    } catch (error) {
-      return { success: false, message: formatErrorMessageCRUD('쿠폰', '추가') };
-    }
-  };
-
-  const resetCouponForm = () => {
-    setCouponForm({
-      name: '',
-      code: '',
-      discountType: 'amount',
-      discountValue: 0,
-    });
-    toggleCouponForm();
-  };
-
-  // Form validation helpers - now returns validation results instead of calling notifications
+  // Form validation helpers
   const validateProductForm = (field: keyof ProductForm, value: any): ValidationResult => {
     switch (field) {
       case 'price':
@@ -186,31 +130,9 @@ export function useAdminForm({
     }
   };
 
-  const validateCouponForm = (field: keyof CouponForm, value: any): ValidationResult => {
-    if (field === 'discountValue') {
-      if (couponForm.discountType === 'percentage') {
-        if (value > 100) {
-          return { isValid: false, message: '할인율은 100%를 초과할 수 없습니다' };
-        }
-      } else {
-        if (value > 100000) {
-          return { isValid: false, message: formatExceedErrorMessage('할인 금액', 100000) };
-        }
-      }
-      if (value < 0) {
-        return { isValid: false, message: formatErrorMessageProduct('할인값', 0) };
-      }
-    }
-    return { isValid: true };
-  };
-
   // Form update helpers
   const handleProductFormChange = (updates: Partial<ProductForm>) => {
     setProductForm((prev) => ({ ...prev, ...updates }));
-  };
-
-  const handleCouponFormChange = (updates: Partial<CouponForm>) => {
-    setCouponForm((prev) => ({ ...prev, ...updates }));
   };
 
   return {
@@ -227,18 +149,7 @@ export function useAdminForm({
     handleEditProduct,
     handleProductFormChange,
 
-    // Coupon form state
-    showCouponForm,
-    toggleCouponForm,
-    couponForm,
-    resetCouponForm,
-
-    // Coupon form handlers
-    handleCouponSubmit,
-    handleCouponFormChange,
-
     // Validation
     validateProductForm,
-    validateCouponForm,
   };
 }
