+++ src/shared/components/index.ts
@@ -5,3 +5,4 @@ export * from './Badge';
 export * from './Input';
 export * from './Select';
 export * from './Button';
+export * from './Tab';
